% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusPeaks.R
\name{ConsensusPeaks}
\alias{ConsensusPeaks}
\title{ConsensusPeaks: A wrapper for peak merging functions}
\usage{
ConsensusPeaks(
  GENES = "all",
  PEAKS,
  RNA.OR.DNA = c("rna", "dna"),
  METHOD = c("dpc", "hmm"),
  GTF = NULL,
  DP.RESOLUTION = 50,
  DP.ITERATIONS = 1000,
  DP.WEIGHT.THRESHOLD = 0.2,
  DP.N.SD = 1,
  DP.ALPHA.PRIORS = c(1, 2),
  DP.SEED = 123,
  PLOT.MERGED.PEAKS = F,
  OUTPUT.TAG = "",
  OUTPUTDIR = ".",
  WRITE.OUTPUT = T,
  ANNOTATION = NULL
)
}
\arguments{
\item{GENES}{A character vector of genes to be tested}

\item{PEAKS}{A data frame containing the following columns, and potentially extras, usually found in a BED12 file, base 0 system
\describe{
  \item{chr}{chromosomes, character, same format as those identified in GTF file}
  \item{start}{starting position of the peak, integer. base 0}
  \item{end}{end position of the peak, integer, base 0}
  \item{name}{gene id, character}
  \item{score}{p-value associated with the peak}
  \item{strand}{strand of the gene, only +, -, * are accepted}
  \item{blockCount}{number of segments in the peak, integer}
  \item{blockSizes}{size of segments in the peak, BED12 notation, comma-separated}
  \item{blockStarts}{starting positions of segments, BED12 notation, comma-separated}
  \item{sample}{sample_id of samples, character}
}}

\item{RNA.OR.DNA}{One of 'rna' or 'dna' indicating whether peaks are identified on the transcriptome or genome}

\item{METHOD}{One 'dpc', 'union', 'corces'
\describe{
 \item{dpc}{A GMM is fitted on the peaks using a Dirichlet process prior}
 \item{union}{The union of overlapping peaks is reported}
 \item{coerces}{A 'daisy chain' model described in Coerces et al., 2012, Science, where the most statistically significant peak is chosen}
}}

\item{GTF}{The path to a GTF file, character string.}

\item{DP.RESOLUTION}{If the METHOD is 'dpc', specify the resolution of the provided PEAKS data.frame. This usually corresponds to the bin width chosen when the peaks are called, a positive integer.}

\item{DP.ITERATIONS}{If the METHOD is 'dpc', specify the number of iterations used to fit the GMM, a positive integer.}

\item{DP.WEIGHT.THRESHOLD}{If the METHOD is 'dpc', specify the minimum weight used to threshold the Gaussians. This value is required to be numeric and between 0 and 1.}

\item{DP.N.SD}{If the METHOD is 'dpc', specify the standard deviation of the Gaussians required to identify peak boundaries, numeric value, above 0.}

\item{DP.ALPHA.PRIORS}{If the METHOD is 'dpc', specify the priors (alpha, beta) for the Gamma distribution from which the weight concentration prior is drawn, numeric vector of length 2}

\item{DP.SEED}{A seed for reproducibility, integer above 0.}

\item{PLOT.MERGED.PEAKS}{Either a logical value (TRUE or FALSE) indicating all or none of the merged peaks should be plotted. Otherwise, a character vector of genes whose merged peaks should be plotted.}

\item{OUTPUT.TAG}{A character string added to the names of any output files.}

\item{OUTPUTDIR}{Output directory. If the directory does not exist, ConsensusPeaks will attempt to create the directory.}

\item{WRITE.OUTPUT}{A logical value indicating whether the output table should be written.}
}
\value{
A data frame with the same columns as 'PEAKS' indicating the coordinates of the merged peaks in genomic coordinates. If the method is 'dpc', additional columns with the sample names of the samples will indicate the merged p-value using Fisher's Method.
}
\description{
ConsensusPeaks: A wrapper for peak merging functions
}
\examples{
data.path = system.file("extdata", package = "ConsensusPeaks")
gtf = paste0(data.path, "/test.gtf")
peaks.file = paste0(data.path, "/peaks.bed")
peaks = read.table(peaks.file, header = F, sep = '\t', stringsAsFactors = F)
colnames(peaks) = c("chr", "start", "end", "name", "score", "strand", "thickStart", "thickEnd", "itemRGB", "blockCount", "blockSizes", "blockStarts", "sample")

results = ConsensusPeaks(
GENES = c("ENSGXX", "ENSGYY"),
PEAKS = peaks,
RNA.OR.DNA = "rna",
METHOD = "dpc",
GTF = gtf,
DP.RESOLUTION = 10,
DP.ITERATIONS = 1000,
DP.WEIGHT.THRESHOLD = 0.1,
DP.N.SD = 1.5,
DP.ALPHA.PRIORS = c(0.5,5),
DP.SEED = 123,
PLOT.MERGED.PEAKS = F,
OUTPUT.TAG = "TEST",
OUTPUTDIR = ".",
WRITE.OUTPUT = F)

}
