% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed6tobed12.R
\name{.bed6tobed12}
\alias{.bed6tobed12}
\title{Creates a BED12 file from a BED6 file using unique columns of ID.COLS}
\usage{
.bed6tobed12(MERGED.PEAKS, ID.COLS)
}
\arguments{
\item{MERGED.PEAKS}{A GRanges Object, in BED6 format for the merged peaks}

\item{ID.COLS}{Column names of columns whose unique values are used to merge peaks}
}
\value{
A BED12 file of merged peaks
\describe{
  \item{chr}{chromosomes, same as in GTF file}
  \item{start}{starting position of the peak, base 0}
  \item{end}{end position of the peak, base 0}
  \item{name}{gene id}
  \item{score}{p-value associated with the peak}
  \item{strand}{strand of the gene}
  \item{blockCount}{number of segments in the peak}
  \item{blockSizes}{size of segments in the peak, BED12 notation}
  \item{blockStarts}{starting positions of segments, BED12 notation}
}
}
\description{
Creates a BED12 file from a BED6 file using unique columns of ID.COLS
}
