% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.peaks.from.gmm.R
\name{.generate.peaks.from.gmm}
\alias{.generate.peaks.from.gmm}
\title{generate.peaks.from.gmm is a function that takes the output of a GMM and generates peaks}
\usage{
.generate.peaks.from.gmm(dp, PARAMETERS, GENEINFO)
}
\arguments{
\item{PARAMETERS}{A PARAMETERS list with the parameters indicated in the DPDE4PM function}

\item{GENEINFO}{A list of gene attributes generated by the get.gene.anno function}

\item{dp_data}{Results from fitting a Dirichlet Process GMM model}
}
\value{
\describe{
 \item{merged.peaks.filtered.rna}{A merged peaks GenomicRanges object in RNA coordinates}
 \item{merged.peaks.filtered.genome}{A merged peaks GenomicRanges object in genomic coordinates}
}
}
\description{
generate.peaks.from.gmm is a function that takes the output of a GMM and generates peaks
}
