% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.gene.anno.R
\name{.get.gene.anno}
\alias{.get.gene.anno}
\title{A function that returns a list of characteristics of the gene annotation}
\usage{
.get.gene.anno(PARAMETERS, ANNOTATION)
}
\arguments{
\item{PARAMETERS}{A PARAMETERS list with the parameters indicated in the DPDE4PM function}

\item{ANNOTATION}{A data frame describing the annotation of genes in the GTF file, generated by the read.gtf function
\describe{
 \item{chr}{chromosome}
 \item{feature}{genomic feature}
 \item{start}{start coordinate, base 1}
 \item{end}{stop coordinate, base 1}
 \item{strand}{strand}
 \item{gene}{gene id in the GTF file}
 \item{transcript}{transcript id in the GTF fiile}
}}
}
\value{
A list of gene characteristics including the following
\describe{
 \item{anno}{annotation of the gene}
 \item{gene}{gene id}
 \item{chr}{chromosome}
 \item{left}{start position of the gene}
 \item{right}{end position of the gene}
 \item{DNA2RNA}{a structure for converting the DNA to RNA}
 \item{RNA2DNA}{a structure for converting the RNA to DNA}
 \item{dna_length}{distance in bps between start and end genomic coordinates}
 \item{exome_length}{length of region covered by exons}
}
}
\description{
A function that returns a list of characteristics of the gene annotation
}
