% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.p.R
\name{.merge.p}
\alias{.merge.p}
\title{A function used to merge p-values using Fisher's method}
\usage{
.merge.p(PEAKSGR, MERGED.PEAKS, ANNOTATION, PARAMETERS, ID.COLS)
}
\arguments{
\item{PEAKSGR}{A GRanges object, output of the retrieve.peaks.as.granges function}

\item{MERGED.PEAKS}{A GRanges object, output of the generate.peaks.from.gmm function}

\item{ANNOTATION}{A data frame describing the annotation of genes in the GTF file, generated by the read.gtf function
\describe{
 \item{chr}{chromosome}
 \item{feature}{genomic feature}
 \item{start}{start coordinate, base 1}
 \item{end}{stop coordinate, base 1}
 \item{strand}{strand}
 \item{gene}{gene id in the GTF file}
 \item{transcript}{transcript id in the GTF file}
}}

\item{PARAMETERS}{A PARAMETERS list with the parameters indicated in the DPDE4PM function}

\item{ID.COLS}{Column names of columns whose unique values are used to merge peaks}
}
\value{
A p-value matrix with columns fo each unique sample a column with the peak id
}
\description{
A function used to merge p-values using Fisher's method
}
