% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.merged.peaks.R
\name{.plot.merged.peaks}
\alias{.plot.merged.peaks}
\title{Makes a plot of the merged peaks and the fitted curve}
\usage{
.plot.merged.peaks(plotting.data, plot.merged.peaks, PARAMETERS)
}
\arguments{
\item{plot.merged.peaks}{A GRanges object containing the RNA coordinates of the merged peaks in BED6 format}

\item{PARAMETERS}{A PARAMETERS list with the parameters indicated in the DPDE4PM function}
}
\description{
Makes a plot of the merged peaks and the fitted curve
}
