% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.peaks.as.granges.R
\name{.retrieve.peaks.as.granges}
\alias{.retrieve.peaks.as.granges}
\title{Converts the PEAKS BED12 format to a usable GRanges BED6 format for downstream analysis}
\usage{
.retrieve.peaks.as.granges(PEAKS, GENE, DF = F)
}
\arguments{
\item{PEAKS}{A data frame containing the following columns, and potentially extras, usually found in a BED12 file, base 0 system}

\item{GENE}{The ID of the gene}

\item{DF}{Return as a data.frame. Otherwise, return as a GRanges object}
}
\value{
a data.frame or GRanges object in BED6 format containing only peaks assigned to a gene with an extra column for the peak index from where it was extracted and a sample id
}
\description{
Converts the PEAKS BED12 format to a usable GRanges BED6 format for downstream analysis
}
