% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.coverage.hist.R
\name{gene.coverage.hist}
\alias{gene.coverage.hist}
\title{Plots the gene coverage for a given gene}
\usage{
gene.coverage.hist(PEAKS, GENE, ANNOTATION)
}
\arguments{
\item{PEAKS}{A data frame containing the following columns, and potentially extras, usually found in a BED12 file, base 0 system}

\item{ANNOTATION}{Annotations returned from \link[ConsensusPeaks]{read.gtf}.}

\item{GENES}{A character vector of genes to be tested}
}
\value{
ggplot of the gene coverage histogram
}
\description{
Plots the gene coverage for a given gene
}
\examples{
gtf.path = system.file("extdata/test.gtf", package = "ConsensusPeaks")
PARAMETERS = list("GTF" = gtf, "GENE" = "ENSGXX")
annotation = read.gtf(PARAMETERS)
ensgxx = simulate.gaussian.peaks(
  MU = c(100, 150),
  SD = c(10, 20),
  EXTEND.WIDTH = c(50, 25),
  NSAMPLES = c(10, 30),
  GENE = "ENSGXX",
  GTF = gtf,
  ANNOTATION = annotation,
  SEED = 123)
gene.coverage.hist(ensgxx, "ENSGXX", annotation)
}
