% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Return a default value if a variable is null}
\usage{
a \%||\% b
}
\arguments{
\item{a}{The possibly null value}

\item{b}{The default value to return if `a` is null}
}
\value{
`a` if `a` is non-null, `b` if `a` is null
}
\description{
Return a default value if a variable is null
}
\examples{
x <- list(a = 1)
y <- x$b \%||\% 2 # y = 2
z <- x$a \%||\% 10 # z = 1
}
