% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.segmentation.ftc.R
\name{local.minmax}
\alias{local.minmax}
\title{Finds the local minima m and maxima M such that
m_1 < M_1 < m_2 < M_2 < ... < M_{K - 1} < m_{k}}
\usage{
local.minmax(x, threshold = 0)
}
\description{
Finds the local minima m and maxima M such that
m_1 < M_1 < m_2 < M_2 < ... < M_{K - 1} < m_{k}
}
