% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gene.peaks.R
\name{plot.gene.peaks}
\alias{plot.gene.peaks}
\title{Creates a pile up of bed files using the Sushi R package}
\usage{
\method{plot}{gene.peaks}(
  GENE,
  PEAKS,
  GTF = NULL,
  ANNOTATION = NULL,
  OUTPUTDIR = ".",
  OUTPUT.TAG = "",
  PLOT = T
)
}
\arguments{
\item{GENE}{A 'character' gene id corresponding to the gene_id's found in the GTF files and the 'name' column in the peak files}

\item{PEAKS}{data frame containing the following columns, and potentially extras, usually found in a BED12 file, base 0 system
\describe{
  \item{chr}{chromosomes, same as in GTF file}
  \item{start}{starting position of the peak, base 0}
  \item{end}{end position of the peak, base 0}
  \item{name}{gene id}
  \item{score}{p-value associated with the peak}
  \item{strand}{strand of the gene}
  \item{blockCount}{number of segments in the peak}
  \item{blockSizes}{size of segments in the peak, BED12 notation}
  \item{blockStarts}{starting positions of segments, BED12 notation}
  \item{sample}{sample_id of samples}
}}

\item{GTF}{The GTF file used to generate the peaks. This is used to determine the genomic coordinates of the gene.}

\item{OUTPUTDIR}{Output directory}

\item{OUTPUT.TAG}{A character string indicating a tag to track the generated files}

\item{PLOT}{Binary T or F indicating whether to save plot or just return plot}
}
\description{
Creates a pile up of bed files using the Sushi R package
}
