% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gtf.R
\name{read.gtf}
\alias{read.gtf}
\title{read.gtf is a function that imports a gtf file and creates an annotation data frame, adapted from exomePeak}
\usage{
read.gtf(PARAMETERS)
}
\arguments{
\item{PARAMETERS}{A PARAMETERS list with the parameters indicated in the DPDE4PM function}
}
\value{
\describe{
 \item{chr}{chromosome}
 \item{feature}{genomic feature}
 \item{start}{start coordinate, base 1}
 \item{end}{stop coordinate, base 1}
 \item{strand}{strand}
 \item{gene}{gene id in the GTF file}
 \item{transcript}{transcript id in the GTF fiile}
}
}
\description{
read.gtf is a function that imports a gtf file and creates an annotation data frame, adapted from exomePeak
}
