% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.and.fit.R
\name{segment.and.fit}
\alias{segment.and.fit}
\title{Segment and fit a gene}
\usage{
segment.and.fit(
  gene,
  annotation,
  peaks,
  all.samples,
  output.tag,
  output.dir,
  plot.merged.peaks,
  diagnostic,
  truncated.models = FALSE,
  uniform.peak.threshold = 0.75,
  uniform.peak.stepsize = 5,
  histogram.metric = c("jaccard", "intersection", "ks", "mse", "chisq"),
  eps = 1,
  remove.low.entropy = T,
  max.uniform = T,
  seed = 123
)
}
\arguments{
\item{gene}{gene id from gtf file or annotation (if rna) and peaks file used to identify the set of associated peaks to merge}

\item{annotation}{A data frame generated using the read.gtf function for a gtf file}

\item{peaks}{A data frame containing the following columns, and potentially extras, usually found in a BED12 file, base 0 system}

\item{all.samples}{A list of all samples with peaks in the dataset}

\item{output.tag}{A character string added to the names of any output files.}

\item{output.dir}{Output directory. If the directory does not exist, ConsensusPeaks will attempt to create the directory.}

\item{plot.merged.peaks}{Only if the method parameter is set to 'sf'. Either a logical value (TRUE or FALSE) indicating all or none of the merged peaks should be plotted. Otherwise, a character vector of genes whose merged peaks should be plotted.}

\item{diagnostic}{Only if the method parameter is set to 'sf'. A logical value indicating whether diagnostic plots for fitted distributions should be plotted.}

\item{fit.mixtures}{Only if the method parameter is set to 'sf'. A character vector indicating distributions to fit.}

\item{trim.peak.threshold}{See ConsensusPeaks}

\item{trim.peak.stepsize}{See ConsensusPeaks}

\item{residual.tolerance}{See ConsensusPeaks}
}
\description{
Segment and fit a gene
}
