% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.gaussian.peaks.R
\name{simulate.gaussian.peaks}
\alias{simulate.gaussian.peaks}
\title{simulate.gaussian.peaks}
\usage{
\method{simulate}{gaussian.peaks}(
  MU = 1,
  SD = 1,
  EXTEND.WIDTH = 50,
  NSAMPLES = 10,
  GENE,
  GTF = NULL,
  ANNOTATION = NULL,
  SEED = 123
)
}
\arguments{
\item{MU}{An integer vector of means used to simulate a mixture of Gaussians. Values represent the mean bp, counted from the starting position of the transcript coordinate.}

\item{SD}{An integer vector of standard deviations used to simulate a mixture of Gaussians. Units are in bp.}

\item{EXTEND.WIDTH}{An integer vector of widths of peaks in bp.}

\item{NSAMPLES}{An integer vector}

\item{GENE}{The name of a gene in the GTF file on which to simulate peaks}

\item{GTF}{The path to a GTF file used to simulate RNA peaks}

\item{SEED}{A seed for reproducibility}
}
\value{
A data.frame in BED12 format, with an extra 'sample' column which can be used as input to ConsistentPeaks, PEAKS input
}
\description{
simulate.gaussian.peaks
}
\examples{
gtf.path = system.file("extdata", package = "ConsensusPeaks")
gtf = paste0(gtf.path, "/test.gtf")
simulate.gaussian.peaks(
MU = c(100, 150),
SD = c(10, 20),
EXTEND.WIDTH = c(50, 25),
NSAMPLES = c(10, 30),
GENE = "ENSGXX",
GTF = gtf,
SEED = 123
)

}
