% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.and.fit.R
\name{segment.and.fit}
\alias{segment.and.fit}
\title{Segment and fit a gene}
\usage{
segment.and.fit(GENE, PARAMETERS, ANNOTATION, PEAKS)
}
\arguments{
\item{GENE}{}

\item{PEAKS}{}
}
\description{
Segment and fit a gene
}
